+++ packages/msgraph/src/Exceptions/GraphRateLimitException.php
@@ -4,4 +4,16 @@ declare(strict_types=1);
 
 namespace Moox\Msgraph\Exceptions;
 
-class GraphRateLimitException extends GraphException {}
+use Throwable;
+
+class GraphRateLimitException extends GraphException
+{
+    public function __construct(
+        string $message = '',
+        int $code = 0,
+        ?Throwable $previous = null,
+        public readonly ?int $retryAfterSeconds = null,
+    ) {
+        parent::__construct($message, $code, $previous);
+    }
+}